modified

+++ Tourism/Tourism/Models/StayResponse.cs
@@ -10,7 +10,7 @@ namespace Tourism.Models
         public int Id { get; set; }
         public int DestinationId { get; set; }
         public int WhereToStayId { get; set; }
-        public int Kilometer { get; set; }
+        public double Kilometer { get; set; }
         public string Name { get; set; }
         public string Phone1 { get; set; }
         public string Phone2 { get; set; }
